+++ src/LanguageMapOfStrings.php
@@ -1,18 +1,18 @@
-<?php declare(strict_types=1);
+<?php declare(strict_types = 1);
 
 namespace JSKOS;
 
 use InvalidArgumentException;
 
 /**
- * Language map of strings.
+ * Language map of strings for prefLabel.
  */
 class LanguageMapOfStrings extends LanguageMap
 {
     protected static function checkMember($value)
     {
-        if (is_string($value)) {
-            return $value;
+        if (is_scalar($value)) {
+            return "$value";
         } else {
             throw new InvalidArgumentException(
                 'JSKOS\LanguageMapOfStrings may only contain strings'
@@ -23,7 +23,7 @@ class LanguageMapOfStrings extends LanguageMap
     /**
      * Return a data structure to serialize this container as JSON.
      */
-    public function jsonSerializeRoot($context=self::DEFAULT_CONTEXT)
+    public function jsonLDSerialize(string $context = self::DEFAULT_CONTEXT)
     {
         return (object)$this->members;
     }