+++ src/Mapping.php
@@ -1,11 +1,11 @@
-<?php declare(strict_types=1);
+<?php declare(strict_types = 1);
 
 namespace JSKOS;
 
 use JSKOS\Item;
 
 /**
- * A JSKOS %Mapping.
+ * A JSKOS Mapping.
  *
  * @see https://gbv.github.io/jskos/jskos.html#concept-mappings
  */