+++ src/Registry.php
@@ -1,55 +1,55 @@
-<?php declare(strict_types=1);
+<?php declare(strict_types = 1);
 
 namespace JSKOS;
 
 use JSKOS\Item;
 
 /**
- * A JSKOS %Registry.
+ * A JSKOS Registry.
  *
  * @see https://gbv.github.io/jskos/jskos.html#registries
  */
 class Registry extends Item
 {
     const FIELDS = [
-        'concepts'     => ['Set','Access'],
-        'schemes'      => ['Set','Access'],
-        'types'        => ['Set','Access'],
-        'mappings'     => ['Set','Access'],
-        'registries'   => ['Set','Access'],
-        'concordances' => ['Set','Access'],
+        'concepts'     => ['Set', 'AccessPoint'],
+        'schemes'      => ['Set', 'AccessPoint'],
+        'types'        => ['Set', 'AccessPoint'],
+        'mappings'     => ['Set', 'AccessPoint'],
+        'registries'   => ['Set', 'AccessPoint'],
+        'concordances' => ['Set', 'AccessPoint'],
         'extent'       => 'string',
         'languages'    => ['Listing'],
-        'license'      => ['Set','Concept'],
+        'license'      => ['Set', 'Concept'],
     ];
 
     /**
-     * Accesses to concepts schemes in this registry.
+     * AccessPoints to concepts schemes in this registry.
      */
     protected $concepts;
 
     /**
-     * Accesses to concept schemes in this registry.
+     * AccessPoints to concept schemes in this registry.
      */
     protected $schemes;
 
     /**
-     * Accesses to types in this registry.
+     * AccessPoints to types in this registry.
      */
     protected $types;
 
     /**
-     * Accesses to concordances in this registry.
+     * AccessPoints to concordances in this registry.
      */
     protected $concordances;
 
     /**
-     * Accesses to mappings in this registry.
+     * AccessPoints to mappings in this registry.
      */
     protected $mappings;
 
     /**
-     * Accesses to registries in this registry.
+     * AccessPoints to registries in this registry.
      */
     protected $registries;