+++ tests/ConceptSchemeTest.php
@@ -1,4 +1,4 @@
-<?php declare(strict_types=1);
+<?php declare(strict_types = 1);
 
 namespace JSKOS;
 
@@ -11,7 +11,7 @@ class ConceptSchemeTest extends \PHPUnit\Framework\TestCase
     public function testEmptyScheme()
     {
         $scheme = new ConceptScheme();
-        $scheme->types = [new Access(['download'=>'http://example.org/types'])];
+        $scheme->types = [new AccessPoint(['download'=>'http://example.org/types'])];
         $expect = [
             '@context' => 'https://gbv.github.io/jskos/context.json',
             'type'     => ['http://www.w3.org/2004/02/skos/core#ConceptScheme'],